WIP: Remove Affix, add simple TCP relay
This commit changes the nodemanager to use aaaaalbert/repy-doodles's
simple TCP relay for NAT traversal instead of the Affix framework.

If on a RFC1918 "private" IPv4 address, set up port forwarding through
a TCP relay. The relay is transparent from the perspective of a
remote host connecting in, the nodemanager just advertises the
relay's IP and relaying listen port. In the commit version, the
nodemanager also prefixes its name with "NAT$". This is useful for
accounting for NAT nodes on the clearinghouse only, and can be removed
to test with legacy clients like `seash`.

I also fixed what I think was an issue with advertising the node name.
If I understood correctly, a change in node IP address would not
reliably trigger a change in the advertised string. I could be wrong
though.